requirements.txt
Django==4.1.5
Pillow==9.4.0
social-auth-app-django==5.0.0
django-extensions==3.2.1
Werkzeug==2.2.2
pyOpenSSL==23.0.0
django-dotenv==1.4.2
certifi==2022.12.7
easy-thumbnails==2.8.5
redis==4.4.2
django_debug_toolbar==3.8.1
django-jalali-date==1.0.1